refactor(dashnote): align src/ layout and document SDK calls
Move non-SDK utilities (logger, notesCache, rememberedIdentity) out of
src/dash/ into src/lib/, move useMediaQuery into src/hooks/, and add
"SDK method:" JSDoc headers to each src/dash/ operation file so the
folder reads as a tour of Platform calls.
