+++ example-apps/dashnote/src/hooks/useMediaQuery.ts
