+++ example-apps/dashnote/src/lib/logger.ts
